Updates to warped landmark node handling
When a reference landmark node is selected, a cloned copy is created for the PCA warped view. This landmark node has the labels hidden by default and takes the glyph properties from the reference landmark node. It is no longer hidden from the editor so the user can adjust viewing properties.