Adds fcsv as a save option for any markup type
This makes it appear in the Export As context menu
and in the Save Data dialog.

As discussed with the SlicerMorph team, when used in the
Export As context menu it will not set the default
storage node, so that .mrk.json will be the default when
saving a scene or mrb.

Proper operation in the Save Data dialog depends on this
change to the Slicer core:
https://github.com/Slicer/Slicer/pull/5025