Add MarkupEditor
Adds a module to select markup fiducials using a markup
curve node and sets their selection state
when you are done placing the curve points.

If we like the way this works it could be the basis
for other edit operations like splitting lists, cut/copy/paste
between lists, etc.