Remove ResampleCurveWorld call
This was used to ensure a smooth sampling
of the curve node, but is not needed since
we are now using linear edges only.

Plus this led to error conditions for
very small fiducial sets (like features
on a small mammal tooth).

If ever re-enabled it should be made
adaptive to the data size somehow.