Enable multiple VolumeProperty actions
Previously, a VolumeProperty action would set the
target propery with the start values and end values
outside of the action time window, effectively preventing
a sequence of multiple volume propery changes.

With this change you have the option to unclamp before
and/or after the action, allowing you to string together
multiple actions in a sequence.

Note that you can use the end VolumeProperty of one action
as the start for another to provide a smooth transition.

Also this change gives the Volume Properties unique names
to make them easier to work with.