fix(dialog) make other files appear in list
Native dialogs can hide files that don't match the selected
filters, even when doing a save operation.  We want to be able
to see the files that do not match.

Here we switch to the Qt dialogs which do show all files in
the folder even if they don't match.

Non-native dialogs are also easier for consistency in documentation
and tutorials.