Only override of volume rendering pref on first run
This commit adds a setting to track the state of overriding
the volume rendering preferences.

Before, we always forced the pref settings at the start of
each run.

Now the preference override only happens the first time the
preferences are set, and after that a user's selection is preserved
if they have chosen a different option.