ImageStacks: make default setting of image spacing more robust
By default do not set any spacing but set it to 0.0.
If the user does not modify the default spacing then an error is displayed when attempting to load the image.
By default, assume isotropic spacing (a button is added to toggle between isotropic/anisotropic).
If user clicks "auto-detect" button then spacing is read from the image metadata. If 2D images are read (not NRRD) then a warning is displayed. The user can click "don't show again" in the warning popup.

fixes #390